feat: bl without readable stream
Now you can get a `BufferList` without the `readable-stream` bits for a smaller browser bundle.

`bl` now exports 2 classes:

1. `BufferList`
2. `BufferListStream`

`BufferListStream` **is the default export** (and so is backwards compatible). i.e.

```js
const { BufferListStream } = require('bl')
// equivalent to:
const BufferListStream = require('bl')
```

I've updated the README to recommend typical use like:

```js
const { BufferList } = require('bl')
// or
const { BufferListStream } = require('bl')
```

Simply `require('bl/BufferList')` in the browser to get **JUST** the `BufferList` class that does not inherit from `Duplex`.

Added static function `BufferList.isBufferList(obj)` which checks if the passed object has a `_isBufferList` property (similar to how `Buffer.isBuffer(obj)` works).

resolves https://github.com/rvagg/bl/issues/48
resolves https://github.com/rvagg/bl/issues/70
resolves https://github.com/rvagg/bl/issues/71
